+++ packages/jupyter-ai/src/index.ts
@@ -4,7 +4,11 @@ import {
   ILayoutRestorer
 } from '@jupyterlab/application';
 
-import { IWidgetTracker, ReactWidget } from '@jupyterlab/apputils';
+import {
+  IWidgetTracker,
+  ReactWidget,
+  IThemeManager
+} from '@jupyterlab/apputils';
 import { IDocumentWidget } from '@jupyterlab/docregistry';
 import { IGlobalAwareness } from '@jupyter/collaboration';
 import type { Awareness } from 'y-protocols/awareness';
@@ -12,6 +16,8 @@ import { buildChatSidebar } from './widgets/chat-sidebar';
 import { SelectionWatcher } from './selection-watcher';
 import { ChatHandler } from './chat_handler';
 import { buildErrorWidget } from './widgets/chat-error';
+import { completionPlugin } from './completions';
+import { statusItemPlugin } from './status';
 
 export type DocumentTracker = IWidgetTracker<IDocumentWidget>;
 
@@ -21,11 +27,12 @@ export type DocumentTracker = IWidgetTracker<IDocumentWidget>;
 const plugin: JupyterFrontEndPlugin<void> = {
   id: 'jupyter_ai:plugin',
   autoStart: true,
-  optional: [IGlobalAwareness, ILayoutRestorer],
+  optional: [IGlobalAwareness, ILayoutRestorer, IThemeManager],
   activate: async (
     app: JupyterFrontEnd,
     globalAwareness: Awareness | null,
-    restorer: ILayoutRestorer | null
+    restorer: ILayoutRestorer | null,
+    themeManager: IThemeManager | null
   ) => {
     /**
      * Initialize selection watcher singleton
@@ -43,10 +50,11 @@ const plugin: JupyterFrontEndPlugin<void> = {
       chatWidget = buildChatSidebar(
         selectionWatcher,
         chatHandler,
-        globalAwareness
+        globalAwareness,
+        themeManager
       );
     } catch (e) {
-      chatWidget = buildErrorWidget();
+      chatWidget = buildErrorWidget(themeManager);
     }
 
     /**
@@ -60,4 +68,4 @@ const plugin: JupyterFrontEndPlugin<void> = {
   }
 };
 
-export default plugin;
+export default [plugin, statusItemPlugin, completionPlugin];